cCurPosition: Renamed "SetEqual" to "SetEqualToMe" in an attempt to remove the ambiguity as who is getting set equal to who (is the caller setting the callee equal to the caller, or visa versa?)
cRow: Elliminated confusing "RowID" variable, which only added confusion to the task of trying to understand the row being worked on.

cSquare: Similiarily, cSquare no longer contains RowID and ColId, but simply Row and Col.

Both Row and Col are zero-based, so their legal values are 0-8.

cValidation: I had used the cValidation code in _2dPuzzleGenerator.sln as a guide to "whip" this code out.  It had some logic problems, so I cleaned up some its problems.

Program.cs:  Stubbed in a short, sweet bit of code to create a puzzle, and if successful, then print the puzzle.  (This will eventually be "Write the puzzle to a database, so it won't be lost, and will not need to be generated again!"